requirements.txt
customtkinter~=5.2.2
setuptools~=75.3.0
pillow~=11.0.0
tkcalendar~=1.6.1
reportlab~=4.2.5
matplotlib~=3.9.2
pip~=24.3.1
wheel~=0.45.0
pyinstaller~=6.11.1
openpyxl~=3.2.0b1
numpy~=2.1.3
Jinja2~=3.1.4
weasyprint~=63.0
pyparsing~=3.2.0